pyproject.toml
[tool.poetry]
name = "advertising-api"
version = "1.0.0"
description = "API for out-of-home advertising system"
authors = ["brunosarti83"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
sqlmodel = "^0.0.22"
supabase = "^2.9.1"
python-dotenv = "^1.0.1"
slowapi = "^0.1.9"
redis = "^5.0.8"
pydantic = "^2.9.2"
alembic = "^1.13.3"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"
ruff = "^0.6.8"
pylint = "^3.3.1"
mypy = "^1.11.2"
graphene = "^3.3.0"

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"